Fixed a bug where the visualization was not rendering a purple expander node.
Incorrect tree node id numbering was causing the same id to be used for more than on node. This resulted in generation of bad data for Cytoscape which in turned caused the visualization to render incorrectly. The bad logic was using the database item id number instead of a uniquely generated kid id  number. This went undetected for a long time until a new database with small database item id numbers collided with a kid id.